+++ great_expectations/cli/init.py
@@ -29,6 +29,10 @@ from great_expectations.exceptions import (
     DatasourceInitializationError,
 )
 
+from great_expectations.core.usage_statistics.usage_statistics import (
+    send_usage_message,
+)
+
 try:
     from sqlalchemy.exc import SQLAlchemyError
 except ImportError:
@@ -96,6 +100,7 @@ def init(target_directory, view, usage_stats):
 
         try:
             context = DataContext.create(target_directory, usage_statistics_enabled=usage_stats)
+            send_usage_message(data_context=context, event="cli.init.ctx_new", success=True)
         except DataContextError as e:
             # TODO ensure this is covered by a test
             cli_message("<red>{}</red>".format(e))